+++ tests/ViewRenderer/RenderCombinations/view.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * @var string $name
+ */
+
+echo 'test';